+++ code/src/nsleep.c
@@ -1,8 +1,8 @@
 #include "util/nsleep.h"
 
-void nsleep(long n) {
-    long s = n / 1000000000L;
-    long ns = n % 1000000000L;
+void nsleep(long long n) {
+    long long s = n / 1000000000L;
+    long long ns = n % 1000000000L;
 
     const struct timespec delay = {s, ns};
